Practical-Exam-Database-Using-Chatbot/chatbot/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.0
pycryptodome==3.11.0
pymongo==4.7.1